+++ highs/ipm/hipo/auxiliary/Auxiliary.h
@@ -11,37 +11,37 @@
 
 namespace hipo {
 
-void counts2Ptr(std::vector<Int>& ptr, std::vector<Int>& w);
-void inversePerm(const std::vector<Int>& perm, std::vector<Int>& iperm);
-void subtreeSize(const std::vector<Int>& parent, std::vector<Int>& sizes);
-void transpose(const std::vector<Int>& ptr, const std::vector<Int>& rows,
-               std::vector<Int>& ptrT, std::vector<Int>& rowsT);
-void transpose(const std::vector<Int>& ptr, const std::vector<Int>& rows,
-               const std::vector<double>& val, std::vector<Int>& ptrT,
-               std::vector<Int>& rowsT, std::vector<double>& valT);
-void childrenLinkedList(const std::vector<Int>& parent, std::vector<Int>& head,
-                        std::vector<Int>& next);
-void reverseLinkedList(std::vector<Int>& head, std::vector<Int>& next);
-void dfsPostorder(Int node, Int& start, std::vector<Int>& head,
-                  const std::vector<Int>& next, std::vector<Int>& order);
-void processEdge(Int j, Int i, const std::vector<Int>& first,
-                 std::vector<Int>& maxfirst, std::vector<Int>& delta,
-                 std::vector<Int>& prevleaf, std::vector<Int>& ancestor);
-double getDiagStart(Int n, Int k, Int nb, Int n_blocks, std::vector<Int>& start,
-                    bool triang = false);
+void counts2Ptr(std::vector<Int64>& ptr, std::vector<Int64>& w);
+void inversePerm(const std::vector<Int64>& perm, std::vector<Int64>& iperm);
+void subtreeSize(const std::vector<Int64>& parent, std::vector<Int64>& sizes);
+void transpose(const std::vector<Int64>& ptr, const std::vector<Int64>& rows,
+               std::vector<Int64>& ptrT, std::vector<Int64>& rowsT);
+void transpose(const std::vector<Int64>& ptr, const std::vector<Int64>& rows,
+               const std::vector<double>& val, std::vector<Int64>& ptrT,
+               std::vector<Int64>& rowsT, std::vector<double>& valT);
+void childrenLinkedList(const std::vector<Int64>& parent,
+                        std::vector<Int64>& head, std::vector<Int64>& next);
+void reverseLinkedList(std::vector<Int64>& head, std::vector<Int64>& next);
+void dfsPostorder(Int64 node, Int64& start, std::vector<Int64>& head,
+                  const std::vector<Int64>& next, std::vector<Int64>& order);
+void processEdge(Int64 j, Int64 i, const std::vector<Int64>& first,
+                 std::vector<Int64>& maxfirst, std::vector<Int64>& delta,
+                 std::vector<Int64>& prevleaf, std::vector<Int64>& ancestor);
+Int64 getDiagStart(Int64 n, Int64 k, Int64 nb, Int64 n_blocks,
+                   std::vector<Int64>& start, bool triang = false);
 
 template <typename T>
-void permuteVector(std::vector<T>& v, const std::vector<Int>& perm) {
+void permuteVector(std::vector<T>& v, const std::vector<Int64>& perm) {
   // Permute vector v according to permutation perm.
   std::vector<T> temp_v(v);
-  for (Int i = 0; i < v.size(); ++i) v[i] = temp_v[perm[i]];
+  for (Int64 i = 0; i < v.size(); ++i) v[i] = temp_v[perm[i]];
 }
 
 template <typename T>
-void permuteVectorInverse(std::vector<T>& v, const std::vector<Int>& iperm) {
+void permuteVectorInverse(std::vector<T>& v, const std::vector<Int64>& iperm) {
   // Permute vector v according to inverse permutation iperm.
   std::vector<T> temp_v(v);
-  for (Int i = 0; i < v.size(); ++i) v[iperm[i]] = temp_v[i];
+  for (Int64 i = 0; i < v.size(); ++i) v[iperm[i]] = temp_v[i];
 }
 
 template <typename T>
