+++ highs/ipm/hipo/factorhighs/Factorise.h
@@ -13,22 +13,22 @@ namespace hipo {
 class Factorise {
  public:
   // matrix to factorise
-  std::vector<Int> rowsA_{};
-  std::vector<Int> ptrA_{};
+  std::vector<Int64> rowsA_{};
+  std::vector<Int64> ptrA_{};
   std::vector<double> valA_{};
-  Int n_{};
-  Int nzA_{};
+  Int64 n_{};
+  Int64 nzA_{};
 
   // symbolic factorisation
   const Symbolic& S_;
 
   // children in supernodal elimination tree
-  std::vector<Int> first_child_{};
-  std::vector<Int> next_child_{};
+  std::vector<Int64> first_child_{};
+  std::vector<Int64> next_child_{};
 
   // reverse linked lists of chidlren
-  std::vector<Int> first_child_reverse_{};
-  std::vector<Int> next_child_reverse_{};
+  std::vector<Int64> first_child_reverse_{};
+  std::vector<Int64> next_child_reverse_{};
 
   // generated elements, aka Schur complements.
   std::vector<std::vector<double>> schur_contribution_{};
@@ -39,7 +39,7 @@ class Factorise {
   std::vector<std::vector<double>>& sn_columns_;
 
   // swaps of columns for each supernode, ordered locally within a block
-  std::vector<std::vector<Int>> swaps_{};
+  std::vector<std::vector<Int64>> swaps_{};
 
   // Information about 2x2 pivots.
   // If pivot_2x2[sn][i] == 0, 1x1 pivot was used.
@@ -65,12 +65,12 @@ class Factorise {
   DataCollector& data_;
 
  public:
-  void permute(const std::vector<Int>& iperm);
-  void processSupernode(Int sn);
+  void permute(const std::vector<Int64>& iperm);
+  void processSupernode(Int64 sn);
 
  public:
-  Factorise(const Symbolic& S, const std::vector<Int>& rowsA,
-            const std::vector<Int>& ptrA, const std::vector<double>& valA,
+  Factorise(const Symbolic& S, const std::vector<Int64>& rowsA,
+            const std::vector<Int64>& ptrA, const std::vector<double>& valA,
             const Regul& regul, const Log* log, DataCollector& data,
             std::vector<std::vector<double>>& sn_columns);
